[SP-3631] extract InMemoryStorageMock to its own file
