+++ Example/ConsentViewController_ExampleTests/LocalStorage/GDPRUserDefaultsSpec.swift
@@ -12,42 +12,6 @@ import Nimble
 
 // swiftlint:disable force_cast function_body_length
 
-class InMemoryStorageMock: Storage {
-    var storage = [String: Any]()
-
-    func string(forKey defaultName: String) -> String? {
-        return storage[defaultName] as? String
-    }
-
-    func object<T>(ofType type: T.Type, forKey defaultName: String) -> T? where T: Decodable {
-        return storage[defaultName] as? T
-    }
-
-    func set(_ value: Any?, forKey defaultName: String) {
-        storage[defaultName] = value
-    }
-
-    func setObject<T>(_ value: T, forKey defaultName: String) where T: Encodable {
-        storage[defaultName] = value
-    }
-
-    func setValuesForKeys(_ keyedValues: [String: Any]) {
-        keyedValues.forEach { storage[$0.key] = $0.value }
-    }
-
-    func removeObject(forKey defaultName: String) {
-        storage.removeValue(forKey: defaultName)
-    }
-
-    func removeObjects(forKeys keys: [String]) {
-        keys.forEach { storage.removeValue(forKey: $0) }
-    }
-
-    func dictionaryRepresentation() -> [String: Any] {
-        return storage
-    }
-}
-
 class GDPRUserDefaultsSpec: QuickSpec {
     func randomUserConsents() -> GDPRUserConsent {
         return GDPRUserConsent(acceptedVendors: [],
